+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This is the changelog for Cldr v2.19.0 released on November 22nd, 2020.  For old
 
 * Add `Cldr.Locale.timezone_from_locale/{1,2}` to extract a timezone ID from a language tag
 
-* Add option `:style` to `Cldr.quote/2`. This option is either `:word` or `:sentence`.  `:sentence` is the default. Using the locale `en` as an example, the differences in formatting are:
+* Add option `:format` to `Cldr.ellipsis/2`. This option is either `:word` or `:sentence`.  `:sentence` is the default. Using the locale `en` as an example, the differences in formatting are:
 
 ```elixir
 # Default style: :sentence
@@ -20,7 +20,7 @@ iex> Cldr.ellipsis ["And furthermore", "there is much to be done"], locale: "ja"
 iex> Cldr.ellipsis "And furthermore", style: :word
 "And furthermore …"
 
-iex> Cldr.ellipsis ["And furthermore", "there is much to be done"], locale: "ja", style: :word
+iex> Cldr.ellipsis ["And furthermore", "there is much to be done"], locale: "ja", format: :word
 "And furthermore … there is much to be done"
 ```